RDKAAMP-508: [AAMP] refactoring - consolidate lAampCfgValue with iAampCfgValue
Reason for change: simplify configuration management
Test Procedure: Regression testing w/ configuration previously under long group
Risks: Medium

(cherry picked from commit 80ad93f55079276c4dc4906c2d13ea1866d95bd8)
(cherry picked from commit b6c07dd82774337041c2ef3e8489590adb5888a8)

+++ AampCurlStore.h
@@ -273,9 +273,9 @@ struct CurlProgressCbContext
 
 	long long downloadStartTime;
 	long long downloadUpdatedTime;
-	long startTimeout;
-	long stallTimeout;
-	long lowBWTimeout;
+	int startTimeout;
+	int stallTimeout;
+	int lowBWTimeout;
 	double downloadSize;
 	CurlAbortReason abortReason;
 	double downloadNow;
